refactor: convert server adapter to TS module
