+++ index.html
@@ -14,7 +14,6 @@
         if (el) el.textContent = 'App failed to load: ' + (e.message || 'Unknown error');
       });
     </script>
-    <script src="server-adapter.js"></script>
     <script type="module" src="/src/main.ts" defer></script>
   </body>
 </html>
