+++ src/server/index.ts
@@ -1,0 +1,2 @@
+export * from './adapter';
+export { default } from './adapter';
