+++ vite.config.ts
@@ -4,4 +4,12 @@ import { defineConfig } from 'vite';
 export default defineConfig({
   base: '/',
   resolve: { alias: { '@': resolve(__dirname, 'src') } },
+  build: {
+    rollupOptions: {
+      input: {
+        main: resolve(__dirname, 'index.html'),
+        adapter: resolve(__dirname, 'src/server/adapter.ts'),
+      },
+    },
+  },
 });
